# Conflicts:
#	project/src/main/java/com/example/project/ProjectApplication.java
#	project/src/main/java/com/example/project/configuration/security/SecurityConfig.java
#	project/src/main/java/com/example/project/controller/AuthController.java
#	project/src/main/java/com/example/project/controller/UserController.java
#	project/src/main/java/com/example/project/service/IUserService.java
#	project/src/main/java/com/example/project/service/impl/RefreshTokenServiceImpl.java
#	project/src/main/java/com/example/project/service/impl/RoleServiceImpl.java
#	project/src/main/java/com/example/project/service/impl/UserServiceImpl.java
#	project/src/main/resources/application-local.properties

+++ project/src/main/java/com/example/project/controller/UserController.java
@@ -1,14 +1,19 @@
 package com.example.project.controller;
 
-import com.example.project.constants.ErrorMessage;
-import com.example.project.dto.UserCreateDTO;
-import com.example.project.dto.UserDTO;
+import com.example.project.dto.request.UserCreateRequestDTO;
+import com.example.project.dto.response.UserDTO;
 import com.example.project.exception.ResponseMessage;
+import com.example.project.model.Role;
 import com.example.project.model.User;
-import com.example.project.payload.response.UserCreateResponse;
-import com.example.project.payload.response.UserResponse;
+import com.example.project.dto.response.UserCreateResponse;
+import com.example.project.dto.response.UserResponse;
+import com.example.project.repository.IRoleRepository;
+import com.example.project.repository.IUserRepository;
+import com.example.project.service.ErrorService;
 import com.example.project.service.IUserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -19,52 +24,55 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
+
+import static java.net.HttpURLConnection.*;
 
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("api/users")
 public class UserController {
     private final IUserService userService;
-
+    private  final IRoleRepository iRoleRepository;
+    private final IUserRepository iUserRepository;
+    private final MessageSource messageSource;
+    private final ErrorService errorService;
     // Get list user
     @GetMapping("") // http://localhost:3000/api/users?page=1&limit=5
     public ResponseEntity<ResponseMessage<Page<UserDTO>>> getUsers(@RequestParam(value = "keySearch", required = false) String keySearch, @RequestParam("page") int page, @RequestParam("limit") int limit) {
-        Pageable pageable = PageRequest.of(page, limit,Sort.by("createdAt").descending());
+        Pageable pageable = PageRequest.of(page, limit,Sort.by("createdAt").ascending());
         return ResponseEntity.status(HttpStatus.OK)
-                .body(new ResponseMessage<Page<UserDTO>>(ErrorMessage.USER_GET_ALL_SUCCESS, ErrorMessage.ACCOUNT_SUCCESS_CODE,
+                .body(new ResponseMessage<Page<UserDTO>>( errorService.processUserGetAllSuccess(),HTTP_OK,
                         userService.getAllUsers(keySearch, pageable)));
     }
 
     // create user
     @PostMapping("")
-    public ResponseEntity<?> createUser(@RequestBody @Valid UserCreateDTO request, BindingResult result) {
-        try {
-            if (result.hasErrors()) {
-                List<String> errMessage = result.getAllErrors()
-                        .stream()
-                        .map(fieldError -> fieldError.getDefaultMessage())
-                        .toList();
-                return ResponseEntity.badRequest().body(errMessage);
-            }
-            UserDTO newUser = userService.createUser(request);
-            var userModel = UserCreateResponse.builder()
-                    .firstName(newUser.getFirstName())
-                    .lastName(newUser.getLastName())
-                    .email(newUser.getEmail())
-                    .roleId(newUser.getRoleId())
-                    .address(newUser.getAddress())
-                    .phone(newUser.getPhone())
-                    .build();
-            return ResponseEntity.ok(userModel);
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(e.getMessage());
+    public ResponseEntity<ResponseMessage<UserDTO>>  createUser(@RequestBody UserCreateRequestDTO request) throws Exception {
+        if(iUserRepository.existsByEmail(request.getEmail())){
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(errorService.processErrorEmailAlreadyExist(), HTTP_NOT_FOUND));
+        }
+        UserDTO addedUser = userService.createUser(request);
+        if (addedUser != null) {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(errorService.processUserRegisterSuccess(),HTTP_OK, addedUser));
+        } else {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(errorService.processUserRegisterFail(), HTTP_SERVER_ERROR));
         }
     }
     // Get single user
     @GetMapping("/{id}")
     public ResponseEntity<?> getUser(@PathVariable Long id) {
         try {
+            Optional<User> getUser = iUserRepository.findById(id);
+            //Check if account has been deleted
+            if(getUser.get().isDeleteFlag())
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage<>(errorService.processUserIsDelete() , HTTP_OK));
             User user = userService.getUserByID(id);
             return ResponseEntity.ok(UserResponse.mapUser(user));
         } catch (Exception e) {
@@ -74,31 +82,56 @@ public class UserController {
     }
     //Update user
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateUser(@Valid @PathVariable Long id, @Valid @RequestBody UserDTO request) {
-//        UserDTO accountUpdated = userService.updateUser(id, request);
-//        if (accountUpdated != null) {
-//            return ResponseEntity.status(HttpStatus.OK)
-//                    .body(new ResponseMessage(ErrorMessage.USER_UPDATE_SUCCESS, ErrorMessage.ACCOUNT_SUCCESS_CODE,accountUpdated));
-//        } else {
-//            return ResponseEntity.status(HttpStatus.OK)
-//                    .body(new ResponseMessage(ErrorMessage.USER_NOT_FOUND, ErrorMessage.RESOURCE_NOT_FOUND_CODE));
-//        }
-        return null;
+    public ResponseEntity<?> updateUser(@Valid @PathVariable Long id, @Valid @RequestBody UserDTO request){
+        Optional<User> optionalOldUser = iUserRepository.findById(id);
+        if (request.getUpdatedAt().equals(optionalOldUser.get().getUpdatedAt()) ){
+            UserDTO userUpdated = userService.updateUser(id, request);
+            if (userUpdated != null) {
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage(errorService.processUpdateSuccess(),HTTP_OK,userUpdated));
+            } else {
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage(errorService.processNotFound(), HTTP_NOT_FOUND));
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(errorService.processUpdateFail(), HTTP_NOT_FOUND));
+        }
     }
     // Delete User
     @DeleteMapping("/{id}")
     public ResponseEntity<ResponseMessage<UserDTO>> deleteUser(@PathVariable Long id) {
         try {
+            Optional<User> accountDelete = iUserRepository.findById(id);
+            //Check if account has been deleted
+            if(accountDelete.get().isDeleteFlag())
+                return ResponseEntity.status(HttpStatus.OK)
+                        .body(new ResponseMessage<>(errorService.processUserIsDelete() , HTTP_OK));
+            //delete account
             userService.deleteUser(id);
             return ResponseEntity.status(HttpStatus.OK)
-                    .body(new ResponseMessage(ErrorMessage.USER_DELETE_SUCCESS, ErrorMessage.ACCOUNT_SUCCESS_CODE));
+                    .body(new ResponseMessage(errorService.processDeleteSuccess(),HTTP_OK));
         }  catch (NumberFormatException e){
             return ResponseEntity.status(HttpStatus.OK)
-                    .body(new ResponseMessage(ErrorMessage.USER_ID_INVALID, ErrorMessage.USER_BAD_REQUEST));
+                    .body(new ResponseMessage(errorService.processIdInvalid(), HTTP_BAD_REQUEST));
         }
         catch (Exception e) {
             return ResponseEntity.status(HttpStatus.OK)
-                    .body(new ResponseMessage(ErrorMessage.USER_NOT_FOUND, ErrorMessage.RESOURCE_NOT_FOUND_CODE));
+                    .body(new ResponseMessage(errorService.processUserDeleteFail(), HTTP_BAD_REQUEST));
         }
     }
+
+    @GetMapping("/getRoles")
+    public ResponseEntity<ResponseMessage<List<Role>>> getAllRole() {
+        List<Role> listRole;
+        try {
+            listRole = iRoleRepository.findAll();
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.OK)
+                    .body(new ResponseMessage(errorService.processRoleGetAllFail(),HTTP_NOT_FOUND));
+        }
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(new ResponseMessage(errorService.processRoleGetAllSuccess(),HTTP_OK ,listRole ));
+    }
+
 }
